Remove rowCount, data, and roleNames from PModUIController
