+++ src/mods/PModLoader.cpp
@@ -1,175 +1,340 @@
 #include "PModLoader.h"
 
-PModLoader::PModLoader(QObject *parent, QStringList ztdList) : QAbstractListModel(parent), m_ztdList(ztdList)
+// Grabs mods from ZTDs and stores them in database
+// TODO: Add any errors to a list of errors to display to user
+// TODO: Add a check to see if mod already exists in database
+// TODO: Add meta.toml file to ztd if it doesn't exist
+// TODO: If meta.toml does not exist, add to list of errors for user
+// TODO: Let user decide if it's a duplicate or not
+void PModUIController::loadModsFromFile(const QStringList &ztdList)
 {
-    m_ztdList = ztdList;
-}
-
-int PModLoader::rowCount(const QModelIndex &parent) const
-{
-    Q_UNUSED(parent);
-    return m_mods_list.size();
-}
-
-// The data method is used to fetch the data for a given index and role
-// for a QAbstractListModel
-QVariant PModLoader::data(const QModelIndex &index, int role) const
-{
-    qDebug () << "Fetching data for index: " << index.row() << ", role: " << role;
-    if (index.isValid() && index.row() >= 0 && index.row() < m_mods_list.length())
+    // Insert mods into database
+    for (const QString &ztd : ztdList)
     {
-        QSharedPointer<PModItem> mod = m_mods_list[index.row()];
-        qDebug() << "Fetching data for mod: " << mod->title();
-        mod->getData(role); // this just translates the role to the data
+        // Check if ztd already exists in database
+        QVector<QSharedPointer<PModItem>> existingMods = searchMods(Operation::Select, "filename", QFileInfo(ztd).fileName());
+        if (!existingMods.isEmpty()) {
+            qDebug() << "ZTD already exists in database: " << QFileInfo(ztd).fileName();
+            continue; // skip mod
+        }
+        
+        // File loading and parsing
+        PFile ztdFile(this, ztd, FileType::Zip); // Load the ztd file
+        PFileData metaData = ztdFile.read("meta.toml"); // get the meta.toml at the root of the ztd (as a PFileData object)
+        bool foundMeta = !metaData.data.isEmpty(); // did we find a meta.toml file in the ztd?
+
+        // load the meta file so we can parse it
+        PConfigMgr meta;
+        if (foundMeta) {
+            // Load the meta file and get the mod data
+            meta.loadConfig(metaData.data);
+        } else {
+            qDebug() << "No meta config found in ztd: " << ztd;
+        }
+
+        // Get a list of the entrypoints in the ztd file. These are always in ucb, uca, ucs, or ai format.
+        QList<PFileData> entryPoints = ztdFile.readAll({"animals/", "scenery/other/"}, {"ucb", "uca", "ucs", "ai"});
+        int contentModsDetected = entryPoints.size();
+
+        // if found meta file and more than 1 entrypoint, then this is 
+        // a content pack. we will call it a collection.
+        // for these kinds of mods, the 'collection' will be the name of the mod
+        // and every 'mod' within the collection will be delisted in the modlist.
+        // instead, they will be show in the infopane.
+        PModItem mod(this);
+
+        bool isCollection = contentModsDetected > 1;
+        bool isSingleMod = contentModsDetected == 1;
+
+        // ------------------------------------------------------------------- Build the base information for the mod
+
+        if (foundMeta) { // meta.toml
+            // Load the meta file and get the mod data
+            toml::table config = toml::parse(metaData.data);
+            mod = buildModFromToml(config, ztd);
+            mod.setCollection(true); // set the collection flag   
+        } else { // no meta file + no entrypoints. user will need to manually configure this mod.
+            // if no meta file and no entrypoints, then this is a misc mod
+            mod = buildDefaultMod(ztd, entryPoints);
+            mod.setCollection(false);           
+        }
+
+        // ------------------------------------------------------------------- Insert file data (size, date, etc.)
+        generateFileData(ztd, mod);
+
+        // ------------------------------------------------------------------- Set the category for the mod
+        QString category = determineModCategory(entryPoints);
+        mod.setCategory(category);
+
+        // ------------------------------------------------------------------- Set the tags for the mod
+        if (isCollection) {
+            QStringList tags = generateTagsFromConfig(meta);
+            mod.setTags(tags);
+        } else if (isSingleMod) {
+            PConfigMgr cfg(nullptr, entryPoints[0]);
+            QStringList tags = generateTagsFromConfig(cfg);
+            mod.setTags(tags);
+        } else {
+            mod.setTags({}); // no tags found
+        }
+
+        // ------------------------------------------------------------------- Set mod status flags
+        mod.setEnabled(true);
+        mod.setSelected(false);
+        mod.setListed(true); 
+
+        // ------------------------------------------------------------------- Set the icon paths for the mod
+        QStringList iconPaths;
+        QList<PModItem> collectionMods;
+        if (isCollection) { // if it's a collection, we need to create mod listings for each mod in the entrypoint list
+            // at this point, no entrypoint mod will have a meta.toml file, so most data will need to be be pulled from ini files
+            collectionMods = buildCollectionMods(entryPoints);
+        } else {
+            // if it's not a collection, then we can just set the icon paths from the entrypoint
+            if (isSingleMod) {
+                PConfigMgr epConfig(nullptr, entryPoints[0]);
+                QStringList epIconPaths = getIconPngPaths(entryPoints[0], ztd);
+                if (!epIconPaths.isEmpty()) {
+                    mod.setIconPaths(epIconPaths);
+                } else {
+                    mod.setIconPaths({});
+                }
+            } else {
+                // if no entrypoints, then set the icon paths to empty
+                mod.setIconPaths({});
+            }
+        }
+
+        insertMod(mod); // insert the mod into the database
+
+        if (collectionMods.size() > 0) {
+            for (const PModItem &collectionMod : collectionMods) {
+                insertMod(collectionMod); // insert the collection mods into the database
+            }
+        }
+
+
+                    // // ---------------------- Get Description, Title -----------------------
+
+                    // if (category == "Animals") {}
+                    //     QString description = config.getValue("cLongHelp", "1033").toString();
+                    //     QString title = config.getValue("cName", "1033").toString();
+                    //     QString category = determineCategory(config);
+                    //     if (category.isEmpty()) {
+                    //         category = "Unknown";
+                    //     }
     }
 
-    // Return empty if the index is invalid
-    qDebug() << "Invalid index: " << index.row() << ", returning empty QVariant";
-    return QVariant();
+    // // close database
+    // db.closeDatabase();
+    qDebug() << "Loaded mods from ZTDs";
 }
 
-// roleNames are a map that gives each enum role a string descriptor that
-// can be called from QML
-QHash<int, QByteArray> PModLoader::roleNames() const
-{
-    qDebug() << "PModLoader::roleNames() called";
-    return PModItem::roleNames();
-}
-
-// --------------- Manage Mod List ------------------
-
-void PModLoader::loadMods()
-{
-    // Placeholder for loading mods; dynamic sqlite query later
-    beginResetModel();
-    // Clear existing mods
-    m_mods_list.clear();
-
-    // Open db
-    PDatabase db;
-    db.openDatabase();
-    // Load mods from ztds
-    db.loadModsFromZTDs(m_ztdList);
-
-    // Get all mods from database
-    QSqlQuery query = db.getAllMods();
-
-    // Iterate through the results and create PModItem objects
-    while (query.next())
-    {
-        qDebug() << "Loading mod from database: " << query.value("title").toString();
-        QSharedPointer<PModItem> mod = db.populateModItem(query);
-        addMod(mod);
+// Determine the category of the mod based on the file extension and path
+// Possible categories are: Building, Scenery, Animals, misc, Unknown
+QString PModUIController::determineCategory(const PFileData &fileData) {
+    if (fileData.size() == 0) {
+        return "Unknown";
     }
 
-    db.closeDatabase();
-    qDebug() << "Loaded mods from database";
-    endResetModel();
+    switch (fileData.ext) {
+        case "ucb":
+            return "Building";
+        case "ucs":
+            return "Scenery";
+        case "uca":
+            return "Animals";
+        case "ai":
+            QString pathParts = fileData.path.split("/");
+            // category is always the first part of the path
+            QString category = pathParts[0];
+            // return proper case for category
+            return category.toUpper().left(1) + category.mid(1).toLower();
+        default:
+            return "Unknown";
+    }
 }
 
-void PModLoader::reloadMod(int index)
-{
-    // Reload mod from database
-    beginResetModel();
-    if (index >= 0 && index < m_mods_list.size())
-    {
-        // qDebug() << "Reloading mod: " << mod->modTitle();
-        PDatabase db;
-        db.openDatabase();
-        QString id = m_mods_list[index]->id();
-        QSharedPointer<PModItem> newMod = db.getModByPk(id);
-        // Update the mod in the list
-        m_mods_list[index] = std::move(newMod);
-        db.closeDatabase();
+QList<PModItem> PModUIController::buildCollectionMods(QList<PFileData> entryPoints, PModItem &mod) {
+    QList<PModItem> collectionMods;
+    for (const PFileData &entryPoint : entryPoints) {
+        PConfigMgr epConfig(nullptr, entryPoint);
+        PModItem epMod;
+        // first, copy some base mod data from the collection mod
+        epMod.setAuthors(mod.authors());
+        epMod.setId(mod.id());
+        epMod.setVersion(mod.version());
+
+        epMod.setCollectionId(mod.id());
+
+        epMod.setFilename(mod.filename());
+        epMod.setCurrentLocation(mod.currentLocation());
+        epMod.setOriginalLocation(mod.currentLocation());
+        epMod.setDisabledLocation("");
+        epMod.setFileSize(mod.fileSize());
+
+        // generate the rest of the data
+        QString epCategory = determineCategory(entryPoint);
+        epMod.setCategory(epCategory);
+        epMod.setTags(generateTagsFromConfig(entryPoint));
+        epMod.setEnabled(true);
+        epMod.setSelected(false);
+        epMod.setListed(false); // since this is a collection item, we don't want to show it in the modlist
+
+        QStringList epIconPaths = getIconPaths(entryPoint, ztd);
+        if (!epIconPaths.isEmpty()) {
+            epMod.setIconPaths(epIconPaths);
+        } else {
+            epMod.setIconsPaths({});
+        }
+
+        // get the description and title from the config file
+        QString epDescription = determineDescription(epConfig, epCategory);
+        QString epTitle = determineTitle(epConfig, epCategory);
+
+        epMod.setDescription(epDescription);
+        epMod.setTitle(epTitle);
+        collectionMods.append(epMod);
     }
-    endResetModel();
+
 }
 
-// Reload mod from database
-void PModLoader::reloadMod(QSharedPointer<PModItem> mod)
-{
-    // Get index from mod list
-    int index = m_mods_list.indexOf(mod);
-    if (index == -1)
-    {
-        qDebug() << "Mod not found in list: " << mod->title();
-        return;
-    }
-    // Reload mod from database
-    qDebug() << "Reloading mod: " << mod->title();
-    reloadMod(index);
-}
+// Determine the category of the mod based on the meta.toml file in the root of the ztd
+PModItem PModUIController::buildModFromToml(PConfigMgr &config) {
+    PModItem mod;
 
-// Updates the mod list based on the property (filter) and search term (value) for filtering and live search
-void PModLoader::updateModList(QString property, QString value) 
-{
-    // Placeholder for updating mod list; dynamic sqlite query later
-    beginResetModel();
-    // Clear existing mods
-    m_mods_list.clear();
+    // values from TOML file
+    mod.setId(config.getValue("mod_id", "").toString());
+    mod.setTitle(config.getValue("name", "").toString());
+    mod.setAuthors(config.getValue("authors", "").toStringList());
+    mod.setDescription(config.getValue("description", "").toString());
+    mod.setVersion(config.getValue("version", "").toString());
+    mod.setLink(config.getValue("link", "").toString());
 
-    // Open db
-    PDatabase db;
-    db.openDatabase();
-
-    QSqlQuery query;
-
-    // Checks on orderBy and searchTerm
-    if (property.isEmpty() && value.isEmpty()) {
-        query = db.getAllMods();
-    }
-    else if (property.isEmpty() && !value.isEmpty()) {
-        query = db.queryMods("title", value);
-    }
-    else {
-        query = db.queryMods(property, value);
-    }
-
-    // Iterate through the results and create PModItem objects
-    while (query.next())
-    {
-        qDebug() << "Loading mod from database: " << query.value("title").toString();
-        QSharedPointer<PModItem> mod = db.populateModItem(query);
-        addMod(mod);
-    }
-    db.closeDatabase();
-    qDebug() << "Updated mod list from database";
-    endResetModel();
-}
-
-// Removes a mod from the list and database
-void PModLoader::removeMod(int index)
-{
-    // remove mod from list
-    beginRemoveRows(QModelIndex(), index, index);
-    m_mods_list.removeAt(index);
-    endRemoveRows();
-}
-
-// Adds a mod to the list and database
-void PModLoader::addMod(QSharedPointer<PModItem> mod)
-{
-    beginInsertRows(QModelIndex(), m_mods_list.size(), m_mods_list.size());
-    qDebug() << "Adding mod to list: " << mod->title();
-    m_mods_list.append(mod);
-    endInsertRows();
-    qDebug() << "New mod count: " << m_mods_list.size();
-}
-
-void PModLoader::replaceMod(QSharedPointer<PModItem> newMod) {
-    int index = m_mods_list.indexOf(newMod);
-    if (index == -1) {
-        for (int i = 0; i < m_mods_list.size(); ++i) {
-            if (m_mods_list[i]->id() == newMod->id()) {
-                index = i;
-                break;
+    // if dependency table exists, then add the dependency id to the mod and add to db
+    QMap<QString, QVariant> depTable = config.getValue("dependencies", "").toMap();
+    if (!depTable.isEmpty()) {
+        for (const auto &dep : depTable) {
+            PDepDataAccess dependency(mod.id(), dep);
+            if (dependency.isValid()) {
+                mod.setDependencyId(dependency.id());
             }
         }
     }
-
-    if (index != -1) {
-        beginResetModel();
-        m_mods_list[index] = newMod;
-        endResetModel();
+    // if no dependency table, then set the dependency id to None
+    else {
+        mod.setDependencyId("None");
     }
+    mod.setDependencyId(config.getValue("dep_id").value_or("None"));
+
+    return mod;
+}
+
+PModItem PModUIController::buildDefaultMod(const QString &ztdPath) {
+    PModItem mod;
+    mod.setId(QUuid::createUuid().toString(QUuid::WithoutBraces));
+    mod.setTitle(QFileInfo(ztdPath).fileName());
+    mod.setAuthors({"Unknown"});
+    mod.setDescription("No description found");
+    mod.setVersion("1.0.0");
+
+    return mod;
+}
+
+void PModUIController::generateFileData(const QString &filePath, PModItem &mod) {
+    QFileInfo fileInfo(filePath);
+    mod.setFilename(fileInfo.fileName());
+    mod.setFileSize(QString::number(fileInfo.size()));
+    mod.setFileDate(fileInfo.lastModified().toString("yyyy-MM-dd hh:mm:ss"));
+    mod.setCurrentLocation(fileInfo.absolutePath());
+    mod.setOriginalLocation(fileInfo.absolutePath());
+    mod.setDisabledLocation("");
+}
+
+QStringList PModUIController::generateTagsFromConfig(PConfigMgr &config) {
+    QStringList tags = config.getAllKeys("member");
+
+    // Clean up the tags; format in proper case
+    for (int i = 0; i < tags.size(); i++) {
+        QString tag = tags[i];
+        tag[0] = tag[0].toUpper();
+        tags[i] = tag;
+    }
+    return tags;
+}
+
+QStringList PModUIController::getIconPngPaths(PConfigMgr &config, const QString &category, const PFile &ztd) {
+    QStringList aniPaths = getIconAniPaths(config, category);
+    QStringList iconPaths = getIconPaths(aniPaths, ztd);
+    return iconPaths;
+}
+
+// TODO: Expand this to include more categories
+QStringList PModUIController::getIconAniPaths(PConfigMgr &config, const QString &category) {
+    QStringList iconAniPaths;
+    if (category == "Animals") {
+        // animals have 1-2 icon ani paths
+        // [m/Icon] and [f/Icon]. Sometimes only one exists.
+        QString aniPathF = config.getValue("f/Icon", "Icon").toString();
+        QString aniPathM = config.getValue("m/Icon", "Icon").toString();
+        if (!aniPathF.isEmpty()) {
+            iconAniPaths.append(aniPathF);
+        }
+        if (!aniPathM.isEmpty()) {
+            iconAniPaths.append(aniPathM);
+        }
+        return iconAniPaths;
+    } else if (category == "Building" || category == "Scenery") {
+        // objects have just 1 icon ani path
+        return {config.getValue("Icon", "Icon").toString()};
+    } else {
+        return QStringList();
+    }
+}
+
+QStringList PModUIController::getIconPaths(const QStringList &aniPaths, PFile &ztd) {
+    QStringList iconPaths;
+
+    for (const QString &aniPath : aniPaths) {
+        // get the ani path and generate the icon path
+        PFileData aniData = ztd.read(aniPath);
+        PConfigMgr aniConfig(nullptr, aniData);
+
+        QString iconPath = buildGraphicPath(aniConfig);
+        iconPaths.append(iconPath);
+    }
+
+    QStringList pngPaths;
+    for (const QString &path : iconPaths) {
+        // get the ani path and generate the icon path
+        QStringList pngs = PGraphics::generateGraphicsAsPng(iconPaths);
+        for (const QString &png : pngs) {
+            pngPaths.append(png);
+        }
+    }
+    return pngPaths;
+}
+
+// ani files provide the directory structure in key/value pairs
+// [animation]
+// dir0 = "animals"
+// dir1 = "ferret"
+// dir2 = "other"
+// animation = "n" or "N"
+// the number of dirN keys isn't guaranteed to be the same for all ani files
+QString PModUIController::buildGraphicPath(PConfigMgr &aniFile) {
+    // clamp limit to 10 directory parts
+    QString iconPath;
+    for (int i = 0; i < 10; i++) {
+        QString dirKey = "dir" + QString::number(i);
+        QString dirValue = aniFile.getValue(dirKey, "animation").toString();
+        if (dirValue.isEmpty()) {
+            break; // no more directories to add
+        }
+        iconPath += dirValue + "/";
+    }
+    // remove the last "/" from the path
+    iconPath.chop(1);
+
+    return iconPath;
 }
